Move UART ID function to its own file
