+++ fujicom/id8250.h
@@ -1,0 +1,8 @@
+extern int port_identify_uart();
+enum {
+  UART_8250 = 0,
+  UART_16450,
+  UART_16550,
+  UART_16550A,
+};
+
